node-hub/dora-parquet-recorder/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dora-parquet-recorder"
version = "0.1.0"
description = "A generic, zero-copy data recorder for dora."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "dora-rs",
    "pyarrow",
    "pandas",
    "numpy",
    "opencv-python"
]

[dependency-groups]
dev = ["pytest >=8.1.1", "ruff >=0.9.1"]

[project.scripts]
dora-parquet-recorder = "dora_parquet_recorder.main:main"

[tool.ruff.lint]
extend-select = [
  "D",    # pydocstyle
  "UP",   # Ruff's UP rule
  "PERF", # Ruff's PERF rule
  "RET",  # Ruff's RET rule
  "RSE",  # Ruff's RSE rule
  "NPY",  # Ruff's NPY rule
  "N",    # Ruff's N rule
  "I",    # Ruff's I rule
]

# Keep your existing select list...
select = ["E", "W", "F", "I", "B", "UP"] 

# ADD THIS LINE to ignore all Documentation (D) warnings
ignore = ["D", "PERF203"]
